+++ pkg/network/message/message_test.go
@@ -5,10 +5,15 @@ import (
 	"testing"
 
 	"github.com/number571/go-peer/pkg/crypto/hashing"
+	"github.com/number571/go-peer/pkg/crypto/puzzle"
 	"github.com/number571/go-peer/pkg/crypto/random"
 	"github.com/number571/go-peer/pkg/encoding"
 	"github.com/number571/go-peer/pkg/payload"
 	testutils "github.com/number571/go-peer/test/_data"
+)
+
+var (
+	_ payload.IPayload = &sInvalidPayload{}
 )
 
 const (
@@ -17,6 +22,20 @@ const (
 	tcNetworkKey = "network_key"
 	tcProof      = 1096
 )
+
+type sInvalidPayload struct{}
+
+func (p *sInvalidPayload) GetHead() uint64 {
+	return 1
+}
+
+func (p *sInvalidPayload) GetBody() []byte {
+	return []byte{}
+}
+
+func (p *sInvalidPayload) ToBytes() []byte {
+	return []byte{123}
+}
 
 func TestMessage(t *testing.T) {
 	t.Parallel()
@@ -48,25 +67,58 @@ func TestMessage(t *testing.T) {
 		return
 	}
 
-	msg1 := LoadMessage(sett, msg.ToBytes())
+	msg1, err := LoadMessage(sett, msg.ToBytes())
+	if err != nil {
+		t.Error(err)
+		return
+	}
 	if !bytes.Equal(msg.GetPayload().ToBytes(), msg1.GetPayload().ToBytes()) {
 		t.Error("load message not equal new message")
 		return
 	}
 
-	if msg := LoadMessage(sett, []byte{1}); msg != nil {
+	msg2, err := LoadMessage(sett, msg.ToString())
+	if err != nil {
+		t.Error(err)
+		return
+	}
+	if !bytes.Equal(msg.GetPayload().ToBytes(), msg2.GetPayload().ToBytes()) {
+		t.Error("load message not equal new message")
+		return
+	}
+
+	msg3 := NewMessage(sett, pld).(*sMessage)
+	msg3.fHash = random.NewStdPRNG().GetBytes(hashing.CSHA256Size)
+	msg3.fProof = puzzle.NewPoWPuzzle(testutils.TCWorkSize).ProofBytes(msg3.fHash)
+	if _, err := LoadMessage(sett, msg3.ToBytes()); err == nil {
+		t.Error("success load with invalid hash")
+		return
+	}
+
+	msg4 := NewMessage(sett, &sInvalidPayload{})
+	if _, err := LoadMessage(sett, msg4.ToBytes()); err == nil {
+		t.Error("success load with invalid payload")
+		return
+	}
+
+	if _, err := LoadMessage(sett, struct{}{}); err == nil {
+		t.Error("success load with unknown type of message")
+		return
+	}
+
+	if _, err := LoadMessage(sett, []byte{1}); err == nil {
 		t.Error("success load incorrect message")
 		return
 	}
 
 	randBytes := random.NewStdPRNG().GetBytes(encoding.CSizeUint64 + hashing.CSHA256Size)
-	if msg := LoadMessage(sett, randBytes); msg != nil {
+	if _, err := LoadMessage(sett, randBytes); err == nil {
 		t.Error("success load incorrect message")
 		return
 	}
 
 	prng := random.NewStdPRNG()
-	if msg := LoadMessage(sett, prng.GetBytes(64)); msg != nil {
+	if _, err := LoadMessage(sett, prng.GetBytes(64)); err == nil {
 		t.Error("success load incorrect message")
 		return
 	}
@@ -78,7 +130,7 @@ func TestMessage(t *testing.T) {
 		},
 		[]byte{},
 	)
-	if msg := LoadMessage(sett, msgBytes); msg != nil {
+	if _, err := LoadMessage(sett, msgBytes); err == nil {
 		t.Error("success load incorrect payload")
 		return
 	}