+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Delete `pkg/errors`: delete package errors -> replace standard "errors" package
 - Delete `pkg/file_system`: delete package file_system -> replace standard "os" package
 - Delete `cmd/hidden_lake`: move network_key param from config main block to settings block
+- Delete `pkg/network/message`: LoadMessage not return error
 
 <!-- ... -->